requirement.txt
asgiref==3.5.0
Django==3.2.10
django-seed==0.3.1
djangorestframework==3.13.1
Faker==11.3.0
mysqlclient==2.1.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
